Detection du son et modification du multiplier

+++ app/src/main/java/com/example/invadermustdie/PersonListAdapter.java
@@ -31,7 +31,7 @@ public class PersonListAdapter extends ArrayAdapter<Score> {
     public View getView(int position, @Nullable View convertView, @NonNull ViewGroup parent) {
         String date = getItem(position).getDate();
         int score = getItem(position).getScore();
-        int multiplier = getItem(position).getMultiplier();
+        double multiplier = getItem(position).getMultiplier();
 
         LayoutInflater inflater = LayoutInflater.from(mContext);
         convertView = inflater.inflate(myRessource, parent, false);
